+++ test/index.js
@@ -293,16 +293,16 @@ describe('Awesome test.', () => {
   // });
 
   it('6.1 TDPoint.get(\'t\')', () => {
-    const point = new TDPoint({ t: 28, fi: 43 });
+    const pointB1 = new TDPoint({ t: 28, fi: 43 });
     const expectedVal = 28;
-    const testedT = point.get('t');
+    const testedT = pointB1.get('t');
 
     assert(expectedVal === testedT, `FUCKUP: testedT is ${testedT}`)
   });
 
   it('6.2 TDPoint.process({ type: \'heating\' })', () => {
-    const point = new TDPoint({ t: 28, fi: 43 });
-    const testedHeatingProcess = point.process({
+    const pointB1 = new TDPoint({ t: 28, fi: 43 });
+    const pointB2 = pointB1.process({
       type: 'heating',
       finalParams: {
         t: 50
@@ -310,8 +310,22 @@ describe('Awesome test.', () => {
     });
     // const tAfter = testedHeatingProcess.get('t');
     const expectedFiAfter = 14.303548169006595;
-    const testedFiAfter = testedHeatingProcess.get('fi');
+    const testedFiAfter = pointB2.get('fi');
+    /*
+    console.log('h end');
+    console.log(pointB2.getHumidity());
 
+    console.log('e end');
+    console.log(pointB2.getEnthalpy());
+
+    console.log('fi end');
+    console.log(pointB2.get('fi'));
+
+    console.log('t end');
+    console.log(pointB2.get('t'));
+
+    console.log(pointB2.processResult);
+    */
     assert(testedFiAfter === expectedFiAfter, `FUCKUP: testedFiAfter is ${testedFiAfter}`)
   });
 
@@ -392,5 +406,113 @@ describe('Awesome test.', () => {
     // console.log(afterCooling.processResult);
 
     assert(testedH === expectedH, `FUCKUP: testedH is ${testedH}`)
+  });
+
+  it('7.1 TDPoint.process({ type: \'adiabatic\' }) t down (lower than tWB)', () => {
+    const pointB1 = new TDPoint({ t: 28, fi: 43 });
+    const pointB6 = pointB1.process({
+      type: 'adiabatic',
+      finalParams: { t: 10 }
+    });
+    const expectedTB6 = 19.030956508469423;
+    const testedTB6 = pointB6.get('t');
+    /*
+    console.log('h end');
+    console.log(pointB6.getHumidity());
+
+    console.log('e end');
+    console.log(pointB6.getEnthalpy());
+
+    console.log('fi end');
+    console.log(pointB6.get('fi'));
+
+    console.log('t end');
+    console.log(pointB6.get('t'));
+
+    // Результат последнего процесса:
+    console.log(pointB6.processResult);
+    */
+    assert(testedTB6 === expectedTB6, `FUCKUP: tested tB6 is ${testedTB6}`)
+  });
+
+  it('>>> 7.2 FAILED! TDPoint.process({ type: \'adiabatic\' }) t up', () => {
+    const pointB1 = new TDPoint({ t: 28, fi: 43 });
+    const pointB7 = pointB1.process({
+      type: 'adiabatic',
+      finalParams: { t: 40 }
+    });
+    const expectedTB7 = 40;
+    const testedTB7 = pointB7.get('t');
+    /*
+    console.log('h end');
+    console.log(pointB7.getHumidity());
+
+    console.log('e end');
+    console.log(pointB7.getEnthalpy());
+
+    console.log('fi end');
+    console.log(pointB7.get('fi'));
+
+    console.log('t end');
+    console.log(pointB7.get('t'));
+
+    // Результат последнего процесса:
+    console.log(pointB7.processResult);
+    */
+    assert(testedTB7 === expectedTB7, `FUCKUP: tested tB7 is ${testedTB7}`)
+  });
+
+  it('7.3 TDPoint.process({ type: \'adiabatic\' }) fi down (higher than fiB1)', () => {
+    const pointB1 = new TDPoint({ t: 28, fi: 43 });
+    const pointB8 = pointB1.process({
+      type: 'adiabatic',
+      finalParams: { fi: 85 }
+    });
+    const expectedTB8 = 21.391231111503785;
+    const testedTB8 = pointB8.get('t');
+    /*
+    console.log('h end');
+    console.log(pointB8.getHumidity());
+
+    console.log('e end');
+    console.log(pointB8.getEnthalpy());
+
+    console.log('fi end');
+    console.log(pointB8.get('fi'));
+
+    console.log('t end');
+    console.log(pointB8.get('t'));
+
+    // Результат последнего процесса:
+    console.log(pointB8.processResult);
+    */
+    assert(testedTB8 === expectedTB8, `FUCKUP: tested tB8 is ${testedTB8}`)
+  });
+
+  it('>>> 7.4 FAILED! TDPoint.process({ type: \'adiabatic\' }) fi down (lower than fiB1)', () => {
+    const pointB1 = new TDPoint({ t: 28, fi: 43 });
+    const pointB9 = pointB1.process({
+      type: 'adiabatic',
+      finalParams: { fi: 25 }
+    });
+    const expectedTB9 = 30.832329523641235;
+    const testedTB9 = pointB9.get('t');
+    /*
+    console.log('h end');
+    console.log(pointB9.getHumidity());
+
+    console.log('e end');
+    console.log(pointB9.getEnthalpy());
+
+    console.log('fi end');
+    console.log(pointB9.get('fi'));
+
+    console.log('t end');
+    console.log(pointB9.get('t'));
+
+    // Результат последнего процесса:
+    console.log(pointB9.processResult);
+    */
+    assert(testedTB9 === expectedTB9, `FUCKUP: tested tB9 is ${testedTB9}`)
   });
 });